requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
pandera==0.18.0
kaleido==0.2.1
